chore: separate app layer from IO layer
