+++ Cargo.toml
@@ -10,3 +10,4 @@ dirs = "4.0.0"
 serde = {version = "1.0.144",  features = ["derive"]}
 serde_json = "1.0.85"
 structopt = "0.3.26"
+tempfile = "3"
